upload_backend/requirements.txt
flask
openpyxl
pillow
pytesseract
pdf2image
